+++ lab1/src/task7/Task7Test.java
@@ -9,33 +9,33 @@ class Task7Test {
 
     @Test
     public void test_1() {
-        int[] actual = new int[]{2, 6, 3, 9, 0, -3};
+        double[] actual = new double[]{2, 6, 3, 9, 0, -3};
         Task7.sort(actual);
-        int[] expected = new int[]{-3, 0, 2, 3, 6, 9};
+        double[] expected = new double[]{-3, 0, 2, 3, 6, 9};
         assertArrayEquals(expected, actual);
     }
 
     @Test
     public void test_2() {
-        int[] actual = new int[]{1, 2, 3, 4, 5, 6};
+        double[] actual = new double[]{1, 2, 3, 4, 5, 6};
         Task7.sort(actual);
-        int[] expected = new int[]{1, 2, 3, 4, 5, 6};
+        double[] expected = new double[]{1, 2, 3, 4, 5, 6};
         assertArrayEquals(expected, actual);
     }
 
     @Test
     public void test_3() {
-        int[] actual = new int[]{6, 5, 4, 3, 2, 1};
+        double[] actual = new double[]{6, 5, 4, 3, 2, 1};
         Task7.sort(actual);
-        int[] expected = new int[]{1, 2, 3, 4, 5, 6};
+        double[] expected = new double[]{1, 2, 3, 4, 5, 6};
         assertArrayEquals(expected, actual);
     }
 
     @Test
     public void test_4() {
-        int[] actual = new int[]{};
+        double[] actual = new double[]{};
         Task7.sort(actual);
-        int[] expected = new int[]{};
+        double[] expected = new double[]{};
         assertArrayEquals(expected, actual);
     }
 }
